[Quant] Separated implementations for quantized & non-quantized tensors for the squeeze function
Summary: This is a PR that address a concern brought up in https://github.com/pytorch/pytorch/issues/54150 regarding opting for separat kernels for quantized & non-quantized tensors, when possible, as opposed to using conditional if-else.

Test plan: Additional testing was not implemented to test this change because test cases in the existing test case already make use of the squeeze function for various backends.

ghstack-source-id: 0268b4d6c732f96b58056773835822b50f5dfd3f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71639

initial